Revert "Revert "Styling for no results from SCC""
This reverts commit 06b1c0e96f9478645629de6e25ecb3147dcaf3c3.
